Rewrite service test for migration

+++ services/repository/src/test/java/org/sagebionetworks/repo/web/service/WikiMigrationServiceTest.java
@@ -23,6 +23,7 @@ import org.sagebionetworks.repo.model.UserGroup;
 import org.sagebionetworks.repo.model.UserGroupDAO;
 import org.sagebionetworks.repo.model.UserInfo;
 import org.sagebionetworks.repo.model.dao.FileHandleDao;
+import org.sagebionetworks.repo.model.dao.V2WikiPageMigrationDao;
 import org.sagebionetworks.repo.model.dao.WikiPageDao;
 import org.sagebionetworks.repo.model.dao.WikiPageKey;
 import org.sagebionetworks.repo.model.file.FileHandle;
@@ -49,6 +50,8 @@ public class WikiMigrationServiceTest {
 	@Autowired
 	private V2WikiPageDao v2wikiPageDAO;
 	@Autowired
+	private V2WikiPageMigrationDao v2WikiPageMigrationDao;
+	@Autowired
 	private UserGroupDAO userGroupDAO;
 	@Autowired
 	private FileHandleDao fileMetadataDao;
@@ -59,14 +62,13 @@ public class WikiMigrationServiceTest {
 
 	String creatorUserGroupId;
 	List<WikiPageKey> toDelete;
-	List<WikiPageKey> toDeleteForErrorTest;
-	S3FileHandle markdown;
+	List<WikiPageKey> toDeleteForParentCase;
 	UserInfo adminUserInfo;
 	
 	@Before
 	public void before() throws NotFoundException {
 		toDelete = new ArrayList<WikiPageKey>(); 
-		toDeleteForErrorTest = new ArrayList<WikiPageKey>();
+		toDeleteForParentCase = new ArrayList<WikiPageKey>();
 		
 		UserGroup userGroup = userGroupDAO.findGroup(AuthorizationConstants.BOOTSTRAP_USER_GROUP_NAME, false);
 		assertNotNull(userGroup);
@@ -74,18 +76,6 @@ public class WikiMigrationServiceTest {
 		assertNotNull(creatorUserGroupId);
 		
 		adminUserInfo = userManager.getUserInfo(AuthorizationConstants.ADMIN_USER_NAME);
-		
-		// Create a file
-		S3FileHandle meta = new S3FileHandle();
-		meta.setBucketName("bucketName");
-		meta.setKey("key");
-		meta.setContentType("content type");
-		meta.setContentSize(123l);
-		meta.setContentMd5("md5");
-		meta.setCreatedBy(creatorUserGroupId);
-		meta.setFileName("file.txt1");
-		meta = fileMetadataDao.createFile(meta);
-		markdown = meta;
 	}
 	
 	@After
@@ -96,11 +86,10 @@ public class WikiMigrationServiceTest {
 				wikiPageDao.delete(id);
 			}
 		}
-		if(v2wikiPageDAO != null && toDelete != null) {
+		// We only want to delete from V2 with the keys of toDelete when its not the special parent case
+		if(v2wikiPageDAO != null && toDelete != null && toDeleteForParentCase.size() == 0) {
 			for(WikiPageKey id: toDelete) {
-				Long rootId = v2wikiPageDAO.getRootWiki(id.getOwnerObjectId(), id.getOwnerObjectType());
-				//V2WikiPage wiki = v2wikiPageDAO.get(id);
-				V2WikiPage wiki = v2wikiPageDAO.get(new WikiPageKey(id.getOwnerObjectId(), id.getOwnerObjectType(), rootId.toString()));
+				V2WikiPage wiki = v2wikiPageDAO.get(id);
 				String markdownHandleId = wiki.getMarkdownFileHandleId();
 				S3FileHandle markdownHandle = (S3FileHandle) fileMetadataDao.get(markdownHandleId);
 				s3Client.deleteObject(markdownHandle.getBucketName(), markdownHandle.getKey());
@@ -108,72 +97,133 @@ public class WikiMigrationServiceTest {
 				v2wikiPageDAO.delete(id);
 			}
 		}
-		if(v2wikiPageDAO != null && toDeleteForErrorTest != null) {
-			for(WikiPageKey id: toDeleteForErrorTest) {
-				v2wikiPageDAO.delete(id);
-				// markdown file handle will be deleted in the next if statement
+		if(v2wikiPageDAO != null && toDeleteForParentCase != null) {
+			for(int i = 0; i < toDeleteForParentCase.size(); i++) {
+				WikiPageKey key = toDeleteForParentCase.get(i);
+				V2WikiPage wiki = v2wikiPageDAO.get(key);
+				String markdownHandleId = wiki.getMarkdownFileHandleId();
+				S3FileHandle markdownHandle = (S3FileHandle) fileMetadataDao.get(markdownHandleId);
+				s3Client.deleteObject(markdownHandle.getBucketName(), markdownHandle.getKey());
+				fileMetadataDao.delete(markdownHandleId);
+				v2wikiPageDAO.delete(key);
 			}
 		}
-		// Delete the file handles
-		if(markdown != null){
-			fileMetadataDao.delete(markdown.getId());
-		}
 	}
-	@Ignore
+	
 	@Test
 	public void testMigrateSomeWikis() throws NotFoundException, IOException {
 		// Create some wiki pages
 		createWikiPages(1, 3);
-	
+		String userName = adminUserInfo.getIndividualGroup().getName();
 		// Migrate some wiki pages
-		PaginatedResults<WikiMigrationResult> results = wikiMigrationService.migrateSomeWikis(adminUserInfo.getIndividualGroup().getName(), 3, 0, "somePath");
+		PaginatedResults<WikiMigrationResult> results = wikiMigrationService.migrateSomeWikis(userName, 3, 0, "somePath");
 		assertNotNull(results);
 		// Limit of 3 should have been returned
 		assertEquals(3, results.getResults().size());
-		// Size of v2 DB should be equal to the number of wikis successfully migrated
+		// Size of the V2 DB should be equal to the number of wikis successfully migrated
+		assertEquals(3, v2wikiPageDAO.getCount());	
+		
+		// Re-migrate the same wiki page.
+		PaginatedResults<WikiMigrationResult> duplicateMigrationResults = wikiMigrationService.migrateSomeWikis(userName, 1, 0, "somePath");
+		assertNotNull(duplicateMigrationResults);
+		assertEquals(1, duplicateMigrationResults.getResults().size());
+		// Size of the V2 DB should still remain 3.
 		assertEquals(3, v2wikiPageDAO.getCount());	
 	}
-	@Ignore
+
 	@Test
-	public void testMigrateSomeWikisWithError() throws NotFoundException, IOException {
-		// Create an error by inserting a wiki with the same id into the V2 DB before migration
+	public void testMigrateSomeWikisWithParents() throws NotFoundException, IOException {
 		String ownerId = "syn1";
 		ObjectType ownerType = ObjectType.ENTITY;
-		// V2 WikiPage already in the V2 DB
-		V2WikiPage page = new V2WikiPage();
-		page.setId("1");
-		page.setCreatedBy(creatorUserGroupId);
-		page.setModifiedBy(creatorUserGroupId);
-		page.setMarkdownFileHandleId(markdown.getId());
-		page.setTitle("title1");
-		page.setAttachmentFileHandleIds(new ArrayList<String>());
+		
+		WikiPage greatAncestor = new WikiPage();
+		greatAncestor.setId("1");
+		greatAncestor.setParentWikiId(null);
+		greatAncestor.setCreatedBy(creatorUserGroupId);
+		greatAncestor.setModifiedBy(creatorUserGroupId);
+		greatAncestor.setMarkdown("markdown1");
+		greatAncestor.setTitle("title1");			
+		greatAncestor.setAttachmentFileHandleIds(new ArrayList<String>());
 		Map<String, FileHandle> fileNameToFileHandleMap = new HashMap<String, FileHandle>();
-		V2WikiPage result = v2wikiPageDAO.create(page, fileNameToFileHandleMap, ownerId, ownerType, new ArrayList<String>());
-		assertNotNull(result);
-		WikiPageKey key = v2wikiPageDAO.lookupWikiKey(result.getId());
-		toDeleteForErrorTest.add(key);
-		assertEquals(1, v2wikiPageDAO.getCount());
-		// WikiPage with same ID being migrated to V2 should not throw an error. It will update on duplicate ID.
-		// Will implement
-		WikiPage page2 = new WikiPage();
-		page2.setId("1");
-		page2.setCreatedBy(creatorUserGroupId);
-		page2.setModifiedBy(creatorUserGroupId);
-		page2.setMarkdown("markdown1");
-		page2.setTitle("title1");
-		page2.setAttachmentFileHandleIds(new ArrayList<String>());
+		greatAncestor = wikiPageDao.create(greatAncestor, fileNameToFileHandleMap, ownerId, ownerType);
+		assertNotNull(greatAncestor);
+		WikiPageKey key = new WikiPageKey(ownerId, ownerType, greatAncestor.getId());
+		
+		WikiPage ancestor = new WikiPage();
+		ancestor.setId("2");
+		ancestor.setParentWikiId("1");
+		ancestor.setCreatedBy(creatorUserGroupId);
+		ancestor.setModifiedBy(creatorUserGroupId);
+		ancestor.setMarkdown("markdown1");
+		ancestor.setTitle("title1");			
+		ancestor.setAttachmentFileHandleIds(new ArrayList<String>());
 		Map<String, FileHandle> fileNameToFileHandleMap2 = new HashMap<String, FileHandle>();
-		WikiPage result2 = wikiPageDao.create(page2, fileNameToFileHandleMap2, ownerId, ownerType);
-		assertNotNull(result2);
-		WikiPageKey key2 = wikiPageDao.lookupWikiKey(result.getId());
+		ancestor = wikiPageDao.create(ancestor, fileNameToFileHandleMap2, ownerId, ownerType);
+		assertNotNull(ancestor);
+		WikiPageKey key2 = new WikiPageKey(ownerId, ownerType, ancestor.getId());
+
+		WikiPage grandparent = new WikiPage();
+		grandparent.setId("3");
+		grandparent.setParentWikiId("2");
+		grandparent.setCreatedBy(creatorUserGroupId);
+		grandparent.setModifiedBy(creatorUserGroupId);
+		grandparent.setMarkdown("markdown1");
+		grandparent.setTitle("title1");			
+		grandparent.setAttachmentFileHandleIds(new ArrayList<String>());
+		Map<String, FileHandle> fileNameToFileHandleMap3 = new HashMap<String, FileHandle>();
+		grandparent = wikiPageDao.create(grandparent, fileNameToFileHandleMap3, ownerId, ownerType);
+		assertNotNull(grandparent);
+		WikiPageKey key3 = new WikiPageKey(ownerId, ownerType, grandparent.getId());
 		
+		WikiPage parent = new WikiPage();
+		parent.setId("4");
+		parent.setParentWikiId("3");
+		parent.setCreatedBy(creatorUserGroupId);
+		parent.setModifiedBy(creatorUserGroupId);
+		parent.setMarkdown("markdown1");
+		parent.setTitle("title1");			
+		parent.setAttachmentFileHandleIds(new ArrayList<String>());
+		Map<String, FileHandle> fileNameToFileHandleMap4 = new HashMap<String, FileHandle>();
+		parent = wikiPageDao.create(parent, fileNameToFileHandleMap4, ownerId, ownerType);
+		assertNotNull(parent);
+		WikiPageKey key4 = new WikiPageKey(ownerId, ownerType, parent.getId());
+		
+		WikiPage child = new WikiPage();
+		child.setId("5");
+		child.setParentWikiId("4");
+		child.setCreatedBy(creatorUserGroupId);
+		child.setModifiedBy(creatorUserGroupId);
+		child.setMarkdown("markdown1");
+		child.setTitle("title1");			
+		child.setAttachmentFileHandleIds(new ArrayList<String>());
+		Map<String, FileHandle> fileNameToFileHandleMap5 = new HashMap<String, FileHandle>();
+		child = wikiPageDao.create(child, fileNameToFileHandleMap5, ownerId, ownerType);
+		assertNotNull(child);
+		WikiPageKey key5 = new WikiPageKey(ownerId, ownerType, child.getId());
+		
+		// Migrate the PARENT first
+		// greatAncestor, ancestor, and grandparent must be migrated first
+		V2WikiPage parentResult = wikiMigrationService.migrate(parent, adminUserInfo);
+		// By migrating parent, the V2 DB should hold four wiki pages
+		assertEquals(4, v2wikiPageDAO.getCount());
 		/*
-		 * Test after implementing "insert or update on duplicate" in the new DAO
-		 * PaginatedResults<WikiMigrationResult> results = wikiMigrationService.migrateSomeWikis(adminUserInfo.getIndividualGroup().getName(), 1, 0, "somePath");	
-		 */
+		assertEquals(true, v2WikiPageMigrationDao.doesParentExist(parent.getParentWikiId()));
+		assertEquals(true, v2WikiPageMigrationDao.doesParentExist(grandparent.getParentWikiId()));
+		assertEquals(true, v2WikiPageMigrationDao.doesParentExist(ancestor.getParentWikiId()));
+		assertEquals(true, v2WikiPageMigrationDao.doesParentExist(child.getParentWikiId()));
+		*/
+		V2WikiPage childResult = wikiMigrationService.migrate(child, adminUserInfo);
+		assertEquals(5, v2wikiPageDAO.getCount());
 		
-		// Cleanup
-		wikiPageDao.delete(key2);
+		// Add keys to list of wikis to delete / start with child to avoid cascade delete
+		toDeleteForParentCase.add(key5);
+		toDeleteForParentCase.add(key4);
+		toDeleteForParentCase.add(key3);
+		toDeleteForParentCase.add(key2);
+		toDeleteForParentCase.add(key);
+
+		// Nothing to clean up in V1 wiki pages, so casacde deleting is okay
+		toDelete.add(key5);
 	}
 	
 	private void createWikiPages(int start, int end) throws NotFoundException {
@@ -187,7 +237,7 @@ public class WikiMigrationServiceTest {
 			page.setCreatedBy(creatorUserGroupId);
 			page.setModifiedBy(creatorUserGroupId);
 			page.setMarkdown("markdown" + i);
-			page.setTitle("title" + i);
+			page.setTitle("title" + i);			
 			page.setAttachmentFileHandleIds(new ArrayList<String>());
 			Map<String, FileHandle> fileNameToFileHandleMap = new HashMap<String, FileHandle>();
 			WikiPage result = wikiPageDao.create(page, fileNameToFileHandleMap, ownerId, ownerType);
@@ -196,4 +246,5 @@ public class WikiMigrationServiceTest {
 			toDelete.add(key);
 		}
 	}
+
 }
